+++ bundles/alpha.model.abft.codegen/src/2d/init.c
@@ -107,7 +107,7 @@ inline double __min_double(double x, double y){
 #define padded_kernel(i,j) padded_kernel[i][j]
 #define patch(i,j) patch[i][j]
 
-void init(long _P, long _N0, long _N1, long _K0, long _K1, long _PK0, long _PK1, double** kernel, double** padded_kernel, double** patch){
+void init(long _P, long _N0, long _N1, long _K0, long _K1, long _PK0, long _PK1, float** kernel, float** padded_kernel, float** patch){
 	///Parameter checking
 	if (!((_P >= 1 && _N0 >= 1 && _N1 >= 1 && _K0 >= 1 && _K1 >= 1 && _PK0 >= 1 && _PK1 >= 1))) {
 		printf("The value of parameters are not valid.\n");
